Shield arc, commands, and more.
I added the .5 delay per player attack, and directional shield blocking, so if you are behind the mob you should be able to attack it even while its shield is up
You can also easily modify how large the shield blocking arc is eg from 180 deg to 270, or as small as 90 etc.
The weapons given in the test arena have also been switched to pre 1.9 pvp.